requirements.txt
telebot
requests
beautifulsoup4
pillow
opencv-python-headless
numpy